Simulate a set of volleyball

vs_simulate_set_theor and vs_simulate_set_mc are convenience functions for vs_simulate_set(..., method = "theoretical") and vs_simulate_set(..., method = "monte carlo") respectively.

vs_simulate_set(
  rates,
  process_model = "phase",
  serving = NA,
  go_to = 25,
  simple = FALSE,
  id = NULL,
  method = "theoretical"
)

vs_simulate_set_mc(...)

vs_simulate_set_theor(...)

Arguments

rates

list: A two-element list, each element of which is a set of rates as returned by vs_estimate_rates. Experimental: for process_model "sideout", the sideout rate component can be a function. This function will be called at each step of the simulation with the parameters:

team_1_score - the score of team 1 at each point in the set so far

team_2_score - the score of team 2 at each point in the set so far

team_1_rotation - the rotation of team 1 at each point in the set so far (rotations are counted relative to the team's starting rotation, which is 1)

team_2_rotation - the rotation of team 2 at each point in the set so far (rotations are counted relative to the team's starting rotation, which is 1)

serving - the serving team 1 or 2 at each point in the set so far

point_won_by - which team won each point in the set so far (this will be NA for the last entry, because that's the current point that hasn't been simulated yet)

outcome - the outcome of each point in the set so far, either "Sideout" or "Breakpoint" if process_model is "sideout", or details TBD if process_model is "phase" The function should return the sideout rate of the receiving team.

process_model

string: either "sideout" or "phase". The "sideout" model uses the estimated sideout rates (in the rates object) directly. The "phase" model breaks play down into different phases (serve, serve receive, etc) and uses the rates associated with those separate phases

serving

logical: if TRUE, team 1 will serve first. If NA, the team serving first will be chosen at random

go_to

integer: the minimum score that must be reached to end the set (typically 25 for indoor volleyball in sets 1 to 4, 15 in set 5, or 21 in beach volleyball)

simple

logical: if TRUE, return simplified output. Only applicable to method "monte carlo". If simple = TRUE, return the team (1 or 2) that won the set. If simple = FALSE, return extra details in a data.frame

id

: an optional value that (if non-NULL) will be returned in the id column of the returned data frame, if simple is FALSE

method

string: the simulation method to use. Either "monte carlo" or "theoretical". Details TBD

...

: parameters as for vs_simulate_set. vs_simulate_set_theor and vs_simulate_set_mc are convenience functions for vs_simulate_set(..., method = "theoretical") and vs_simulate_set(..., method = "monte carlo") respectively

Value

Integer (1 or 2) or a data frame, depending on the value of simple

See also

Examples

if (FALSE) {
  library(datavolley)
  x <- dv_read(dv_example_file())
  rates <- vs_estimate_rates(x, target_team = "each")

  vs_simulate_set(rates) ## simulate a single set
  vs_simulate_match(rates) ## simulate a match
  ## so given the performances of the two teams during that match, we expect
  ##  that the home team should have won, with 3-0 being the most likely scoreline

  ## compare to the actual match result
  summary(x)
}

## sideout rates as a function for team 2
sofun2 <- function(serving, point_won_by, ...) {
    ## if team 2 won their previous sideout opportunity, their sideout rate is 0.6
    ## otherwise it's 0.5
    prevso <- tail(na.omit(point_won_by[serving == 1]), 1)
    if (length(prevso) < 1 || prevso == 1) {
        ## first sideout opportunity or lost the last one
        0.5
    } else {
        0.6
    }
}

rates <- list(list(sideout = 0.55), ## first team has constant 55% sideout rate
               list(sideout = sofun2)) ## function for team 2's sideout rate

## need to use method = "monte carlo" for this
vs_simulate_set(rates = rates, process_model = "sideout", method = "monte carlo")
#>    team_1_score team_2_score team_1_rotation team_2_rotation serving
#> 1             0            0               1               1       2
#> 2             1            0               2               1       1
#> 3             2            0               2               1       1
#> 4             2            1               2               2       2
#> 5             2            2               2               2       2
#> 6             2            3               2               2       2
#> 7             2            4               2               2       2
#> 8             2            5               2               2       2
#> 9             2            6               2               2       2
#> 10            3            6               3               2       1
#> 11            3            7               3               3       2
#> 12            4            7               4               3       1
#> 13            4            8               4               4       2
#> 14            4            9               4               4       2
#> 15            4           10               4               4       2
#> 16            5           10               5               4       1
#> 17            5           11               5               5       2
#> 18            5           12               5               5       2
#> 19            6           12               6               5       1
#> 20            7           12               6               5       1
#> 21            7           13               6               6       2
#> 22            7           14               6               6       2
#> 23            7           15               6               6       2
#> 24            8           15               1               6       1
#> 25            9           15               1               6       1
#> 26            9           16               1               1       2
#> 27            9           17               1               1       2
#> 28            9           18               1               1       2
#> 29           10           18               2               1       1
#> 30           10           19               2               2       2
#> 31           10           20               2               2       2
#> 32           10           21               2               2       2
#> 33           11           21               3               2       1
#> 34           12           21               3               2       1
#> 35           13           21               3               2       1
#> 36           13           22               3               3       2
#> 37           14           22               4               3       1
#> 38           14           23               4               4       2
#> 39           14           24               4               4       2
#> 40           15           24               5               4       1
#> 41           16           24               5               4       1
#>    point_won_by set_won_by    outcome
#> 1             1          2    Sideout
#> 2             1          2 Breakpoint
#> 3             2          2    Sideout
#> 4             2          2 Breakpoint
#> 5             2          2 Breakpoint
#> 6             2          2 Breakpoint
#> 7             2          2 Breakpoint
#> 8             2          2 Breakpoint
#> 9             1          2    Sideout
#> 10            2          2    Sideout
#> 11            1          2    Sideout
#> 12            2          2    Sideout
#> 13            2          2 Breakpoint
#> 14            2          2 Breakpoint
#> 15            1          2    Sideout
#> 16            2          2    Sideout
#> 17            2          2 Breakpoint
#> 18            1          2    Sideout
#> 19            1          2 Breakpoint
#> 20            2          2    Sideout
#> 21            2          2 Breakpoint
#> 22            2          2 Breakpoint
#> 23            1          2    Sideout
#> 24            1          2 Breakpoint
#> 25            2          2    Sideout
#> 26            2          2 Breakpoint
#> 27            2          2 Breakpoint
#> 28            1          2    Sideout
#> 29            2          2    Sideout
#> 30            2          2 Breakpoint
#> 31            2          2 Breakpoint
#> 32            1          2    Sideout
#> 33            1          2 Breakpoint
#> 34            1          2 Breakpoint
#> 35            2          2    Sideout
#> 36            1          2    Sideout
#> 37            2          2    Sideout
#> 38            2          2 Breakpoint
#> 39            1          2    Sideout
#> 40            1          2 Breakpoint
#> 41            2          2    Sideout