Simulate a set of volleyball

Simulate a set of volleyball

vs_simulate_set(
  rates,
  process_model = "phase",
  serving = NA,
  go_to = 25,
  simple = FALSE,
  id = NULL
)

Arguments

rates
list: A two-element list, each element of which is a set of rates as returned by vs_estimate_rates. Experimental: for process_model "sideout", the sideout rate component can be a function. This function will be called at each step of the simulation with the parameters:

team_1_score - the score of team 1 at each point in the set so far

team_2_score - the score of team 2 at each point in the set so far

serving - the serving team 1 or 2 at each point in the set so far

point_won_by - which team won each point in the set so far (this will be NA for the last entry, because that's the current point that hasn't been simulated yet)

outcome - the outcome of each point in the set so far, either "Sideout" or "Breakpoint" if process_model is "sideout", or details TBD if process_model is "phase"

go_to

process_model
string: either "sideout" or "phase". Details TBD

serving
logical: if TRUE, team 1 will serve first. If NA, the team serving first will be chosen at random

go_to
integer: the minimum score that must be reached to end the set (typically 25 for indoor volleyball in sets 1 to 4, 15 in set 5, or 21 in beach volleyball)

simple
logical: if TRUE, just return the team (1 or 2) that won. If FALSE, return extra details in a data.frame

id
: an optional value that (if non-NULL) will be returned in the id column of the returned data frame, if simple is FALSE

Value

Integer (1 or 2) or a data frame, depending on the value of simple

See also

Examples

if (FALSE) {
  library(datavolley)
  x <- dv_read(dv_example_file())
  rates <- list(vs_estimate_rates(x, target_team = home_team(x)),
                vs_estimate_rates(x, target_team = visiting_team(x)))
  vs_simulate_set(rates) ## simulate a single set
  vs_simulate_match(rates, n = 100, simple = TRUE) ## simulate a match 100 times
  ## so given the performances of the two teams during that match, we expect
  ##  that the home team should have won, with 3-0 being the most likely scoreline

  ## compare to the actual match result
  summary(x)
}

## sideout rates as a function for team 2
sofun2 <- function(serving, point_won_by, ...) {
    ## if team 2 won their previous sideout opportunity, their sideout rate is 0.6
    ## otherwise it's 0.5
    prevso <- tail(na.omit(point_won_by[serving == 1]), 1)
    if (length(prevso) < 1 || prevso == 1) {
        ## first sideout opportunity or lost the last one
        0.5
    } else {
        0.6
    }
}

rates <- list(list(sideout = 0.55), ## first team has constant 55% sideout rate
               list(sideout = sofun2)) ## function for team 2's sideout rate

vs_simulate_set(rates = rates, process_model = "sideout")

#>    team_1_score team_2_score serving point_won_by set_won_by    outcome
#> 1             0            0       2            1          2    Sideout
#> 2             1            0       1            1          2 Breakpoint
#> 3             2            0       1            1          2 Breakpoint
#> 4             3            0       1            2          2    Sideout
#> 5             3            1       2            1          2    Sideout
#> 6             4            1       1            2          2    Sideout
#> 7             4            2       2            1          2    Sideout
#> 8             5            2       1            1          2 Breakpoint
#> 9             6            2       1            2          2    Sideout
#> 10            6            3       2            2          2 Breakpoint
#> 11            6            4       2            2          2 Breakpoint
#> 12            6            5       2            1          2    Sideout
#> 13            7            5       1            1          2 Breakpoint
#> 14            8            5       1            2          2    Sideout
#> 15            8            6       2            1          2    Sideout
#> 16            9            6       1            1          2 Breakpoint
#> 17           10            6       1            2          2    Sideout
#> 18           10            7       2            2          2 Breakpoint
#> 19           10            8       2            1          2    Sideout
#> 20           11            8       1            2          2    Sideout
#> 21           11            9       2            2          2 Breakpoint
#> 22           11           10       2            2          2 Breakpoint
#> 23           11           11       2            2          2 Breakpoint
#> 24           11           12       2            1          2    Sideout
#> 25           12           12       1            2          2    Sideout
#> 26           12           13       2            2          2 Breakpoint
#> 27           12           14       2            2          2 Breakpoint
#> 28           12           15       2            1          2    Sideout
#> 29           13           15       1            1          2 Breakpoint
#> 30           14           15       1            2          2    Sideout
#> 31           14           16       2            1          2    Sideout
#> 32           15           16       1            2          2    Sideout
#> 33           15           17       2            2          2 Breakpoint
#> 34           15           18       2            2          2 Breakpoint
#> 35           15           19       2            1          2    Sideout
#> 36           16           19       1            2          2    Sideout
#> 37           16           20       2            2          2 Breakpoint
#> 38           16           21       2            2          2 Breakpoint
#> 39           16           22       2            2          2 Breakpoint
#> 40           16           23       2            1          2    Sideout
#> 41           17           23       1            2          2    Sideout
#> 42           17           24       2            1          2    Sideout
#> 43           18           24       1            2          2    Sideout